Changed FundsListener to a stub in the stub package and fixed issues
